refactor: extract PDF template generation to separate modules (closes #349)
- Created backend/src/services/pdf_templates/ module with:
  - builders.py: All formatting and helper functions for PDFs
  - invoice_template.py: Sales invoice HTML generation
  - purchase_template.py: Purchase invoice HTML generation
  - __init__.py: Public API exports
- Moved 800+ lines of PDF template and formatting code from routes/invoices.py
- Kept _money() and _is_interstate_supply() utilities in invoices.py (used for invoice creation logic)
- Reduced invoices.py from 1760 to 829 lines (931 line reduction)
- All module syntax verified; imports working correctly
- Routes/handlers remain unchanged for backward compatibility
