+++ backend/src/api/routes/invoices.py
@@ -1,5 +1,4 @@
 from io import BytesIO
-from html import escape
 from datetime import date, datetime
 from collections import defaultdict
 
@@ -27,25 +26,26 @@ from src.api.deps import get_active_company, get_current_user
 from src.services.series import generate_next_number
 from src.services.financial_year import get_active_fy, get_fy_for_date
 from src.services.invoice_payments import build_invoice_payment_summaries
+from src.services.pdf_templates import (
+    _build_invoice_html,
+    _build_multi_copy_invoice_html,
+    _copy_label,
+)
 
 router = APIRouter()
+
+
+def _is_interstate_supply(company_gst: str | None, ledger_gst: str | None) -> bool:
+    """Check if the invoice is for an interstate supply based on GST state codes."""
+    if not company_gst or not ledger_gst or len(company_gst) < 2 or len(ledger_gst) < 2:
+        return False
+    return company_gst[:2] != ledger_gst[:2]
 
 
 def _money(value: Decimal) -> Decimal:
     return value.quantize(Decimal("0.01"), rounding=ROUND_HALF_UP)
 
 
-def _is_interstate_supply(company_gst: str | None, ledger_gst: str | None) -> bool:
-    if not company_gst or not ledger_gst or len(company_gst) < 2 or len(ledger_gst) < 2:
-        return False
-    return company_gst[:2] != ledger_gst[:2]
-
-
-def _extract_pan_from_gstin(gstin: str | None) -> str | None:
-    normalized = (gstin or "").strip().upper()
-    if len(normalized) != 15:
-        return None
-    return normalized[2:12]
 
 
 def _generate_next_number(
@@ -679,940 +679,9 @@ def update_invoice(
         raise HTTPException(status_code=500, detail=str(e))
 
 
-def _fmt_currency(value: float, currency_code: str | None = None) -> str:
-    code = currency_code or "USD"
-    try:
-        if code == "INR":
-            return f"\u20b9{value:,.2f}"
-        elif code == "EUR":
-            return f"\u20ac{value:,.2f}"
-        elif code == "GBP":
-            return f"\u00a3{value:,.2f}"
-        else:
-            return f"${value:,.2f}"
-    except Exception:
-        return f"{value:,.2f}"
 
 
-def _e(text: str | None) -> str:
-    return escape(text or "")
 
-
-def _amount_in_words_indian(amount: float, currency_code: str | None = None) -> str:
-    """Convert a monetary amount to words using the Indian numbering system (crores/lakhs)."""
-    _ONES = ['', 'One', 'Two', 'Three', 'Four', 'Five', 'Six', 'Seven',
-             'Eight', 'Nine', 'Ten', 'Eleven', 'Twelve', 'Thirteen',
-             'Fourteen', 'Fifteen', 'Sixteen', 'Seventeen', 'Eighteen', 'Nineteen']
-    _TENS = ['', '', 'Twenty', 'Thirty', 'Forty', 'Fifty', 'Sixty', 'Seventy', 'Eighty', 'Ninety']
-
-    def _two(n: int) -> str:
-        if n < 20:
-            return _ONES[n]
-        return _TENS[n // 10] + (' ' + _ONES[n % 10] if n % 10 else '')
-
-    def _three(n: int) -> str:
-        if n == 0:
-            return ''
-        if n < 100:
-            return _two(n)
-        return _ONES[n // 100] + ' Hundred' + (' ' + _two(n % 100) if n % 100 else '')
-
-    try:
-        amount = round(amount, 2)
-        rupees = int(amount)
-        paise = round((amount - rupees) * 100)
-        currency_label = 'Rupees' if not currency_code or currency_code == 'INR' else currency_code
-
-        if rupees == 0 and paise == 0:
-            return f'Zero {currency_label} Only'
-
-        parts: list[str] = []
-        n = rupees
-        crores = n // 10_000_000
-        n %= 10_000_000
-        lakhs = n // 100_000
-        n %= 100_000
-        thousands = n // 1000
-        remainder = n % 1000
-
-        if crores:
-            parts.append(_three(crores) + ' Crore')
-        if lakhs:
-            parts.append(_two(lakhs) + ' Lakh')
-        if thousands:
-            parts.append(_two(thousands) + ' Thousand')
-        if remainder:
-            parts.append(_three(remainder))
-
-        rupee_words = ' '.join(parts) if parts else 'Zero'
-        result = f'{rupee_words} {currency_label}'
-        if paise:
-            result += f' and {_two(paise)} Paise'
-        return result + ' Only'
-    except Exception:
-        return ''
-
-
-def _fmt_rate(value: float) -> str:
-    return f"{value:.2f}".rstrip("0").rstrip(".")
-
-
-def _build_pdf_tax_header_cells(interstate_supply: bool) -> str:
-    if interstate_supply:
-        return '<th class="right">IGST %</th><th class="right">IGST Amt</th><th class="right">Total Tax</th>'
-    return (
-        '<th class="right">SGST %</th>'
-        '<th class="right">SGST Amt</th>'
-        '<th class="right">CGST %</th>'
-        '<th class="right">CGST Amt</th>'
-        '<th class="right">Total Tax</th>'
-    )
-
-
-def _build_pdf_table_colgroup(interstate_supply: bool) -> str:
-    if interstate_supply:
-        return (
-            '<colgroup>'
-            '<col style="width: 4%;" />'
-            '<col style="width: 20%;" />'
-            '<col style="width: 7%;" />'
-            '<col style="width: 8%;" />'
-            '<col style="width: 6%;" />'
-            '<col style="width: 6%;" />'
-            '<col style="width: 12%;" />'
-            '<col style="width: 7%;" />'
-            '<col style="width: 10%;" />'
-            '<col style="width: 10%;" />'
-            '<col style="width: 10%;" />'
-            '</colgroup>'
-        )
-    return (
-        '<colgroup>'
-        '<col style="width: 3%;" />'
-        '<col style="width: 16%;" />'
-        '<col style="width: 6%;" />'
-        '<col style="width: 7%;" />'
-        '<col style="width: 5%;" />'
-        '<col style="width: 5%;" />'
-        '<col style="width: 10%;" />'
-        '<col style="width: 6%;" />'
-        '<col style="width: 8%;" />'
-        '<col style="width: 6%;" />'
-        '<col style="width: 8%;" />'
-        '<col style="width: 8%;" />'
-        '<col style="width: 12%;" />'
-        '</colgroup>'
-    )
-
-
-def _build_pdf_tax_row_cells(item: InvoiceItem, currency: str, interstate_supply: bool) -> str:
-    gst_rate = float(item.gst_rate or 0)
-    tax_amount = float(item.tax_amount or 0)
-
-    if interstate_supply:
-        igst_amount = float(item.igst_amount or 0)
-        if tax_amount > 0 and igst_amount == 0:
-            igst_amount = tax_amount
-        igst_rate = gst_rate if igst_amount > 0 else 0
-        return (
-            f'<td class="right">{_fmt_rate(igst_rate)}%</td>'
-        f'<td class="right">{_fmt_currency(igst_amount, currency)}</td>'
-            f'<td class="right">{_fmt_currency(tax_amount, currency)}</td>'
-        )
-
-    sgst_amount = float(item.sgst_amount or 0)
-    cgst_amount = float(item.cgst_amount or 0)
-    if tax_amount > 0 and sgst_amount == 0 and cgst_amount == 0:
-      cgst_amount = float(_money(Decimal(str(tax_amount)) / Decimal("2")))
-      sgst_amount = float(_money(Decimal(str(tax_amount)) - Decimal(str(cgst_amount))))
-
-    split_rate = gst_rate / 2 if (tax_amount > 0 or sgst_amount > 0 or cgst_amount > 0) else 0
-    return (
-        f'<td class="right">{_fmt_rate(split_rate)}%</td>'
-      f'<td class="right">{_fmt_currency(sgst_amount, currency)}</td>'
-        f'<td class="right">{_fmt_rate(split_rate)}%</td>'
-      f'<td class="right">{_fmt_currency(cgst_amount, currency)}</td>'
-        f'<td class="right">{_fmt_currency(tax_amount, currency)}</td>'
-    )
-
-
-def _build_pdf_tax_breakup_rows(invoice: Invoice, currency: str) -> str:
-  cgst = float(invoice.cgst_amount or 0)
-  sgst = float(invoice.sgst_amount or 0)
-  igst = float(invoice.igst_amount or 0)
-
-  # GST display rule for PDFs: show either IGST or CGST+SGST.
-  if igst > 0:
-    return f"<p>IGST: {_fmt_currency(igst, currency)}</p>"
-  return (
-    f"<p>CGST: {_fmt_currency(cgst, currency)}</p>"
-    f"<p>SGST: {_fmt_currency(sgst, currency)}</p>"
-  )
-
-
-def _pdf_unit_price_including_tax(item: InvoiceItem) -> float:
-    quantity = Decimal(str(item.quantity or 0))
-    if quantity <= 0:
-        return float(item.unit_price or 0)
-
-    line_total = Decimal(str(item.line_total or 0))
-    return float(_money(line_total / quantity))
-
-
-def _pdf_display_unit(unit: str | None) -> str:
-    normalized_unit = (unit or "Pieces").strip()
-    if not normalized_unit:
-        normalized_unit = "Pieces"
-    if normalized_unit.lower() == "pieces":
-        return "Pcs"
-    return normalized_unit
-
-
-def _pdf_display_quantity(quantity: float | Decimal | int | None, allow_decimal: bool | None) -> str:
-  value = Decimal(str(quantity or 0))
-  if not allow_decimal and value == value.to_integral_value():
-    return str(int(value))
-
-  as_text = format(value, "f").rstrip("0").rstrip(".")
-  return as_text or "0"
-
-
-def _build_pdf_payment_details_html(invoice_bank_accounts: list[CompanyAccount]) -> str:
-    if not invoice_bank_accounts:
-        return '<p class="muted-text">No bank account marked to display on invoice.</p>'
-
-    blocks: list[str] = []
-    for account in invoice_bank_accounts:
-        blocks.append(
-            (
-                '<div class="invoice-sheet__bank-card">'
-                f"<p>Bank: {_e(account.bank_name) or 'N/A'}</p>"
-                f"<p>Branch: {_e(account.branch_name) or 'N/A'}</p>"
-          f"<p>Account Name: {_e(account.account_name) or 'N/A'}</p>"
-                f"<p>A/C No: {_e(account.account_number) or 'N/A'}</p>"
-                f"<p>IFSC: {_e(account.ifsc_code) or 'N/A'}</p>"
-                "</div>"
-            )
-        )
-
-    return f'<div class="invoice-sheet__bank-cards">{"".join(blocks)}</div>'
-
-
-def _build_purchase_invoice_html(invoice: Invoice, products: list[Product]) -> str:
-    """Generate HTML for a purchase invoice (supplier at top-left, your company top-right,
-    no bank details in footer, optional supplier ref row)."""
-    currency = invoice.company_currency_code or "USD"
-    inv_number = invoice.invoice_number or f"#{invoice.id}"
-    inv_date = invoice.invoice_date.strftime("%d %b %Y") if invoice.invoice_date else (invoice.created_at.strftime("%d %b %Y") if invoice.created_at else "N/A")
-    interstate_supply = _is_interstate_supply(invoice.company_gst, invoice.ledger_gst)
-    tax_header_cells = _build_pdf_tax_header_cells(interstate_supply)
-    table_colgroup = _build_pdf_table_colgroup(interstate_supply)
-
-    product_map = {p.id: p for p in products}
-
-    item_rows = ""
-    for idx, item in enumerate(invoice.items or [], start=1):
-        prod = product_map.get(item.product_id)
-        product_name = _e(prod.name) if prod else f"Product #{item.product_id}"
-        item_description = _e(item.description)
-        product_cell_html = product_name
-        if item_description:
-            product_cell_html = f"{product_name}<br><span class=\"muted-text\">{item_description}</span>"
-        sku = _e(prod.sku) if prod else "N/A"
-        hsn = _e(item.hsn_sac or (prod.hsn_sac if prod else None) or "N/A")
-        unit = _e(_pdf_display_unit(getattr(prod, "unit", None) if prod else None))
-        quantity_display = _pdf_display_quantity(item.quantity, getattr(prod, "allow_decimal", None) if prod else None)
-        tax_row_cells = _build_pdf_tax_row_cells(item, currency, interstate_supply)
-
-        item_rows += f"""
-        <tr>
-          <td>{idx}</td>
-          <td>{product_cell_html}</td>
-          <td>{sku}</td>
-          <td>{hsn}</td>
-          <td class="right">{quantity_display}</td>
-          <td>{unit}</td>
-          <td class="right">{_fmt_currency(_pdf_unit_price_including_tax(item), currency)}</td>
-          {tax_row_cells}
-          <td class="right">{_fmt_currency(float(item.line_total), currency)}</td>
-        </tr>"""
-
-    # Supplier details are stored as ledger_* (buyer_* in DB) on purchase invoices
-    supplier_detail_parts = []
-    if invoice.ledger_gst:
-        supplier_detail_parts.append(f"GST: {_e(invoice.ledger_gst)}")
-        supplier_pan = _extract_pan_from_gstin(invoice.ledger_gst)
-        if supplier_pan:
-            supplier_detail_parts.append(f"PAN: {_e(supplier_pan)}")
-    if invoice.ledger_phone:
-        supplier_detail_parts.append(f"Phone: {_e(invoice.ledger_phone)}")
-    supplier_details = " &middot; ".join(supplier_detail_parts)
-
-    # Your company (receiving / bill-to party)
-    company_detail_parts = []
-    if invoice.company_gst:
-        company_detail_parts.append(f"GST: {_e(invoice.company_gst)}")
-        company_pan = _extract_pan_from_gstin(invoice.company_gst)
-        if company_pan:
-            company_detail_parts.append(f"PAN: {_e(company_pan)}")
-    company_details = " &middot; ".join(company_detail_parts)
-
-    # Optional supplier reference row
-    supplier_ref_html = ""
-    if invoice.supplier_invoice_number:
-        supplier_ref_html = f"""
-  <section class="invoice-sheet__supplierref">
-    <span class="eyebrow">Supplier Ref:</span>
-    <span class="invoice-sheet__supplierref-value">{_e(invoice.supplier_invoice_number)}</span>
-  </section>"""
-
-    round_off_amount = float(invoice.round_off_amount or 0)
-    show_round_off = bool(invoice.apply_round_off and round_off_amount != 0)
-    round_off_html = (
-        f'<p>Round off: {_fmt_currency(round_off_amount, currency)}</p>' if show_round_off else ''
-    )
-    tax_breakup_rows = _build_pdf_tax_breakup_rows(invoice, currency)
-
-    html = f"""<!DOCTYPE html>
-<html>
-<head>
-<meta charset="utf-8">
-<style>
-  @page {{
-    size: A4;
-    margin: 15mm 18mm;
-  }}
-  * {{ margin: 0; padding: 0; box-sizing: border-box; }}
-  body {{
-    font-family: -apple-system, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
-    font-size: 10px;
-    color: #1f2937;
-    line-height: 1.5;
-  }}
-  .eyebrow {{
-    font-size: 8px;
-    font-weight: 600;
-    text-transform: uppercase;
-    letter-spacing: 0.08em;
-    color: #6b7280;
-    margin-bottom: 2px;
-  }}
-  .invoice-sheet {{
-    width: 100%;
-  }}
-  .invoice-sheet__header {{
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-    padding-bottom: 16px;
-    border-bottom: 2px solid #e5e7eb;
-    margin-bottom: 12px;
-  }}
-  .invoice-sheet__header h3 {{
-    font-size: 16px;
-    font-weight: 700;
-    margin-bottom: 4px;
-  }}
-  .invoice-sheet__header p {{
-    font-size: 9px;
-    color: #6b7280;
-    margin-bottom: 1px;
-  }}
-  .invoice-sheet__header-right {{
-    text-align: right;
-  }}
-  .invoice-sheet__header-right h3 {{
-    font-size: 14px;
-  }}
-  .invoice-sheet__titleblock {{
-    display: flex;
-    align-items: center;
-    gap: 12px;
-    margin-bottom: 16px;
-    padding: 10px 14px;
-    background: #f9fafb;
-    border: 1px solid #e5e7eb;
-    border-radius: 6px;
-  }}
-  .invoice-sheet__titleblock h2 {{
-    font-size: 14px;
-    font-weight: 700;
-  }}
-  .invoice-sheet__titleblock p {{
-    font-size: 9px;
-    color: #6b7280;
-    margin-left: auto;
-  }}
-  .invoice-badge {{
-    display: inline-block;
-    padding: 3px 10px;
-    border-radius: 4px;
-    font-size: 9px;
-    font-weight: 600;
-    color: #065f46;
-    background: #ecfdf5;
-  }}
-  .invoice-sheet__supplierref {{
-    margin-bottom: 12px;
-    font-size: 9px;
-    color: #374151;
-  }}
-  .invoice-sheet__supplierref-value {{
-    font-size: 9px;
-    font-weight: 600;
-    color: #1f2937;
-    margin-left: 4px;
-  }}
-  .invoice-sheet__table {{
-    width: 100%;
-    table-layout: fixed;
-    border-collapse: collapse;
-    margin-bottom: 16px;
-    font-size: 8px;
-  }}
-  .invoice-sheet__table thead th {{
-    background: #f3f4f6;
-    color: #374151;
-    font-weight: 600;
-    font-size: 7px;
-    text-transform: uppercase;
-    letter-spacing: 0.04em;
-    line-height: 1.2;
-    padding: 5px 4px;
-    border-bottom: 2px solid #d1d5db;
-    text-align: left;
-    white-space: normal;
-    overflow-wrap: anywhere;
-  }}
-  .invoice-sheet__table thead th.right {{
-    text-align: right;
-  }}
-  .invoice-sheet__table tbody td {{
-    padding: 5px 4px;
-    border-bottom: 1px solid #e5e7eb;
-    vertical-align: middle;
-    white-space: normal;
-    overflow-wrap: anywhere;
-  }}
-  .invoice-sheet__table tbody td.right {{
-    text-align: right;
-  }}
-  .invoice-sheet__table tbody tr:last-child td {{
-    border-bottom: 2px solid #d1d5db;
-  }}
-  .invoice-sheet__footer {{
-    display: flex;
-    justify-content: flex-end;
-    margin-top: 8px;
-  }}
-  .invoice-sheet__totals {{
-    text-align: right;
-    min-width: 220px;
-  }}
-  .invoice-sheet__totals p {{
-    font-size: 9px;
-    color: #4b5563;
-    margin-bottom: 1px;
-  }}
-  .invoice-sheet__total-value {{
-    font-size: 20px !important;
-    font-weight: 700;
-    color: #1a56db !important;
-    margin-top: 4px;
-    margin-bottom: 4px;
-  }}
-  .muted-text {{
-    font-size: 8px;
-    color: #9ca3af;
-  }}
-  .invoice-amount-words {{
-    font-size: 8px;
-    font-style: italic;
-    color: #374151;
-    margin-top: 2px;
-  }}
-</style>
-</head>
-<body>
-<div class="invoice-sheet">
-  <header class="invoice-sheet__header">
-    <div>
-      <p class="eyebrow">Supplier</p>
-      <h3>{_e(invoice.ledger_name) or 'Unknown supplier'}</h3>
-      <p>{_e(invoice.ledger_address) or 'Address not provided'}</p>
-      <p>{supplier_details}</p>
-    </div>
-    <div class="invoice-sheet__header-right">
-      <p class="eyebrow">Bill To</p>
-      <h3>{_e(invoice.company_name) or 'Your company'}</h3>
-      <p>{_e(invoice.company_address) or 'Address not provided'}</p>
-      <p>{company_details}</p>
-    </div>
-  </header>
-
-  <div class="invoice-sheet__titleblock">
-    <span class="invoice-badge">Purchase Invoice</span>
-    <h2>Invoice {_e(inv_number)}</h2>
-    <p>Date: {inv_date} &nbsp;&middot;&nbsp; Currency: {_e(currency)}</p>
-  </div>
-{supplier_ref_html}
-  <section>
-    <table class="invoice-sheet__table">
-      {table_colgroup}
-      <thead>
-        <tr>
-          <th>#</th>
-          <th>Item</th>
-          <th>SKU</th>
-          <th>HSN/SAC</th>
-          <th class="right">Qty</th>
-          <th>Unit</th>
-          <th class="right">Unit Price<br><span style="font-size: 6px; font-weight: 500; text-transform: none;">(with tax)</span></th>
-          {tax_header_cells}
-          <th class="right">Amount</th>
-        </tr>
-      </thead>
-      <tbody>
-        {item_rows}
-      </tbody>
-    </table>
-  </section>
-
-  <section class="invoice-sheet__footer">
-    <div class="invoice-sheet__totals">
-      <p class="eyebrow">Tax breakup</p>
-      <p>Taxable: {_fmt_currency(float(invoice.taxable_amount or 0), currency)}</p>
-      {tax_breakup_rows}
-      <p>Total tax: {_fmt_currency(float(invoice.total_tax_amount or 0), currency)}</p>
-      {round_off_html}
-      <p class="eyebrow" style="margin-top: 10px;">Total due</p>
-      <p class="invoice-sheet__total-value">{_fmt_currency(float(invoice.total_amount), currency)}</p>
-      <p class="invoice-amount-words">{_amount_in_words_indian(float(invoice.total_amount), currency)}</p>
-      <p class="muted-text">Received by {_e(invoice.company_name) or 'Your company'}</p>
-    </div>
-  </section>
-</div>
-</body>
-</html>"""
-    return html
-
-
-def _copy_label(n: int) -> str:
-    _labels = {1: "Original", 2: "Duplicate", 3: "Triplicate"}
-    if n in _labels:
-        return _labels[n]
-    suffix = ("th" if (n % 100) in (11, 12, 13) else {1: "st", 2: "nd", 3: "rd"}.get(n % 10, "th"))
-    return f"{n}{suffix} Copy"
-
-
-def _build_invoice_html(invoice: Invoice, products: list[Product], invoice_bank_accounts: list[CompanyAccount] | None = None, copy_label: str = "Original") -> str:
-    if invoice_bank_accounts is None:
-        invoice_bank_accounts = []
-    
-    if invoice.voucher_type == "purchase":
-        return _build_purchase_invoice_html(invoice, products)
-
-    currency = invoice.company_currency_code or "USD"
-    voucher_label = "Sales" if invoice.voucher_type == "sales" else "Purchase"
-    inv_number = invoice.invoice_number or f"#{invoice.id}"
-    inv_date = invoice.invoice_date.strftime("%d %b %Y") if invoice.invoice_date else (invoice.created_at.strftime("%d %b %Y") if invoice.created_at else "N/A")
-    interstate_supply = _is_interstate_supply(invoice.company_gst, invoice.ledger_gst)
-    tax_header_cells = _build_pdf_tax_header_cells(interstate_supply)
-    table_colgroup = _build_pdf_table_colgroup(interstate_supply)
-
-    product_map = {p.id: p for p in products}
-
-    # Build line item rows
-    item_rows = ""
-    for idx, item in enumerate(invoice.items or [], start=1):
-        prod = product_map.get(item.product_id)
-        product_name = _e(prod.name) if prod else f"Product #{item.product_id}"
-        item_description = _e(item.description)
-        product_cell_html = product_name
-        if item_description:
-            product_cell_html = f"{product_name}<br><span class=\"muted-text\">{item_description}</span>"
-        sku = _e(prod.sku) if prod else "N/A"
-        hsn = _e(item.hsn_sac or (prod.hsn_sac if prod else None) or "N/A")
-        unit = _e(_pdf_display_unit(getattr(prod, "unit", None) if prod else None))
-        quantity_display = _pdf_display_quantity(item.quantity, getattr(prod, "allow_decimal", None) if prod else None)
-        tax_row_cells = _build_pdf_tax_row_cells(item, currency, interstate_supply)
-
-        item_rows += f"""
-        <tr>
-          <td>{idx}</td>
-          <td>{product_cell_html}</td>
-          <td>{sku}</td>
-          <td>{hsn}</td>
-          <td class="right">{quantity_display}</td>
-          <td>{unit}</td>
-          <td class="right">{_fmt_currency(_pdf_unit_price_including_tax(item), currency)}</td>
-          {tax_row_cells}
-          <td class="right">{_fmt_currency(float(item.line_total), currency)}</td>
-        </tr>"""
-
-    # Company details
-    company_detail_parts = []
-    if invoice.company_gst:
-        company_detail_parts.append(f"GST: {_e(invoice.company_gst)}")
-        company_pan = _extract_pan_from_gstin(invoice.company_gst)
-        if company_pan:
-            company_detail_parts.append(f"PAN: {_e(company_pan)}")
-    if invoice.company_phone:
-        company_detail_parts.append(f"Phone: {_e(invoice.company_phone)}")
-    company_details = " &middot; ".join(company_detail_parts)
-
-    company_contact_parts = []
-    if invoice.company_email:
-        company_contact_parts.append(f"Email: {_e(invoice.company_email)}")
-    if invoice.company_website:
-        company_contact_parts.append(f"Web: {_e(invoice.company_website)}")
-    company_contact = " &middot; ".join(company_contact_parts)
-
-    # Bill-to details
-    billto_parts = []
-    if invoice.ledger_gst:
-        billto_parts.append(f"GST: {_e(invoice.ledger_gst)}")
-        billto_pan = _extract_pan_from_gstin(invoice.ledger_gst)
-        if billto_pan:
-            billto_parts.append(f"PAN: {_e(billto_pan)}")
-    if invoice.ledger_phone:
-        billto_parts.append(f"Phone: {_e(invoice.ledger_phone)}")
-    billto_details = " &middot; ".join(billto_parts)
-
-    invoice_title = "Tax Invoice" if invoice.ledger_gst else "Sales Invoice"
-
-    round_off_amount = float(invoice.round_off_amount or 0)
-    show_round_off = bool(invoice.apply_round_off and round_off_amount != 0)
-    round_off_html = (
-        f'<p>Round off: {_fmt_currency(round_off_amount, currency)}</p>' if show_round_off else ''
-    )
-    tax_breakup_rows = _build_pdf_tax_breakup_rows(invoice, currency)
-    payment_details_html = _build_pdf_payment_details_html(invoice_bank_accounts)
-    reference_notes_html = ""
-    if invoice.reference_notes:
-        reference_notes_html = f"""
-  <section class=\"invoice-sheet__reference-notes\">
-    <p class=\"eyebrow\">Reference notes</p>
-    <p class=\"invoice-sheet__reference-notes-value\">{_e(invoice.reference_notes)}</p>
-  </section>"""
-
-    html = f"""<!DOCTYPE html>
-<html>
-<head>
-<meta charset="utf-8">
-<style>
-  @page {{
-    size: A4;
-    margin: 15mm 18mm;
-  }}
-  * {{ margin: 0; padding: 0; box-sizing: border-box; }}
-  body {{
-    font-family: -apple-system, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
-    font-size: 10px;
-    color: #1f2937;
-    line-height: 1.5;
-  }}
-  .eyebrow {{
-    font-size: 8px;
-    font-weight: 600;
-    text-transform: uppercase;
-    letter-spacing: 0.08em;
-    color: #6b7280;
-    margin-bottom: 2px;
-  }}
-  .invoice-sheet {{
-    width: 100%;
-  }}
-  .invoice-sheet__header {{
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-    padding-bottom: 16px;
-    border-bottom: 2px solid #e5e7eb;
-    margin-bottom: 16px;
-  }}
-  .invoice-sheet__header h3 {{
-    font-size: 16px;
-    font-weight: 700;
-    margin-bottom: 4px;
-  }}
-  .invoice-sheet__header p {{
-    font-size: 9px;
-    color: #6b7280;
-    margin-bottom: 1px;
-  }}
-  .invoice-sheet__meta {{
-    text-align: right;
-  }}
-  .invoice-badge {{
-    display: inline-block;
-    padding: 3px 10px;
-    border-radius: 4px;
-    font-size: 9px;
-    font-weight: 600;
-    color: #1a56db;
-    background: #eff6ff;
-    margin-bottom: 6px;
-  }}
-  .invoice-sheet__meta h2 {{
-    font-size: 14px;
-    font-weight: 700;
-    margin-bottom: 2px;
-  }}
-  .invoice-sheet__meta p {{
-    font-size: 9px;
-    color: #6b7280;
-  }}
-  .invoice-sheet__billto {{
-    background: #f9fafb;
-    border: 1px solid #e5e7eb;
-    border-radius: 6px;
-    padding: 12px 14px;
-    margin-bottom: 16px;
-  }}
-  .invoice-sheet__billto h4 {{
-    font-size: 12px;
-    font-weight: 600;
-    margin-bottom: 2px;
-  }}
-  .invoice-sheet__billto p {{
-    font-size: 9px;
-    color: #4b5563;
-    margin-bottom: 1px;
-  }}
-  .invoice-sheet__reference-notes {{
-    border: 1px dashed #d1d5db;
-    border-radius: 6px;
-    padding: 10px 14px;
-    margin-bottom: 14px;
-    background: #f9fafb;
-  }}
-  .invoice-sheet__reference-notes-value {{
-    font-size: 9px;
-    color: #374151;
-  }}
-  .invoice-sheet__table {{
-    width: 100%;
-    table-layout: fixed;
-    border-collapse: collapse;
-    margin-bottom: 16px;
-    font-size: 8px;
-  }}
-  .invoice-sheet__table thead th {{
-    background: #f3f4f6;
-    color: #374151;
-    font-weight: 600;
-    font-size: 7px;
-    text-transform: uppercase;
-    letter-spacing: 0.04em;
-    line-height: 1.2;
-    padding: 5px 4px;
-    border-bottom: 2px solid #d1d5db;
-    text-align: left;
-    white-space: normal;
-    overflow-wrap: anywhere;
-  }}
-  .invoice-sheet__table thead th.right {{
-    text-align: right;
-  }}
-  .invoice-sheet__table tbody td {{
-    padding: 5px 4px;
-    border-bottom: 1px solid #e5e7eb;
-    vertical-align: middle;
-    white-space: normal;
-    overflow-wrap: anywhere;
-  }}
-  .invoice-sheet__table tbody td.right {{
-    text-align: right;
-  }}
-  .invoice-sheet__table tbody tr:last-child td {{
-    border-bottom: 2px solid #d1d5db;
-  }}
-  .invoice-sheet__footer {{
-    display: flex;
-    justify-content: flex-end;
-    margin-top: 8px;
-  }}
-  .invoice-sheet__bank-section {{
-    margin-top: 14px;
-    border-top: 1px solid #e5e7eb;
-    padding-top: 10px;
-  }}
-  .invoice-sheet__bank-section p {{
-    font-size: 9px;
-    color: #4b5563;
-    margin-bottom: 1px;
-  }}
-  .invoice-sheet__bank-cards {{
-    display: flex;
-    flex-wrap: wrap;
-    gap: 14px;
-    align-items: stretch;
-  }}
-  .invoice-sheet__bank-card {{
-    flex: 1 1 240px;
-    max-width: 320px;
-    border: 1px solid #e5e7eb;
-    border-radius: 10px;
-    background: #f9fafb;
-    padding: 10px 12px;
-  }}
-  .invoice-sheet__bank-card p {{
-    white-space: normal;
-    overflow-wrap: anywhere;
-    word-break: break-word;
-  }}
-  .invoice-sheet__bank-card-title {{
-    font-size: 10px !important;
-    font-weight: 700;
-    color: #1f2937 !important;
-    margin-bottom: 4px !important;
-  }}
-  .invoice-sheet__totals {{
-    min-width: 240px;
-    text-align: right;
-  }}
-  .invoice-sheet__totals p {{
-    font-size: 9px;
-    color: #4b5563;
-    margin-bottom: 1px;
-  }}
-  .invoice-sheet__total-value {{
-    font-size: 20px !important;
-    font-weight: 700;
-    color: #1a56db !important;
-    margin-top: 4px;
-    margin-bottom: 4px;
-  }}
-  .muted-text {{
-    font-size: 8px;
-    color: #9ca3af;
-  }}
-  .invoice-title {{
-    text-align: center;
-    font-size: 18px;
-    font-weight: 700;
-    letter-spacing: 0.06em;
-    text-transform: uppercase;
-    color: #1f2937;
-    padding-bottom: 10px;
-    border-bottom: 2px solid #e5e7eb;
-    margin-bottom: 16px;
-  }}
-  .copy-label {{
-    text-align: right;
-    font-size: 8px;
-    font-weight: 600;
-    color: #6b7280;
-    text-transform: uppercase;
-    letter-spacing: 0.08em;
-    margin-bottom: 4px;
-  }}
-  .invoice-amount-words {{
-    font-size: 8px;
-    font-style: italic;
-    color: #374151;
-    margin-top: 2px;
-  }}
-</style>
-</head>
-<body>
-<div class="invoice-sheet">
-  <div class="copy-label">{copy_label}</div>
-  <div class="invoice-title">{invoice_title}</div>
-  <header class="invoice-sheet__header">
-    <div>
-      <p class="eyebrow">Billed by</p>
-      <h3>{_e(invoice.company_name) or 'Company not set'}</h3>
-      <p>{_e(invoice.company_address) or 'Address not provided'}</p>
-      <p>{company_details}</p>
-      <p>{company_contact}</p>
-    </div>
-    <div class="invoice-sheet__meta">
-      <span class="invoice-badge">{voucher_label}</span>
-      <h2>Invoice {_e(inv_number)}</h2>
-      <p>Date: {inv_date}</p>
-      <p>Currency: {_e(currency)}</p>
-    </div>
-  </header>
-
-  <section class="invoice-sheet__billto">
-    <p class="eyebrow">Bill to</p>
-    <h4>{_e(invoice.ledger_name) or 'Unknown ledger'}</h4>
-    <p>{_e(invoice.ledger_address) or 'Address not provided'}</p>
-    <p>{billto_details}</p>
-  </section>
-
-  {reference_notes_html}
-
-  <section>
-    <table class="invoice-sheet__table">
-      {table_colgroup}
-      <thead>
-        <tr>
-          <th>#</th>
-          <th>Item</th>
-          <th>SKU</th>
-          <th>HSN/SAC</th>
-          <th class="right">Qty</th>
-          <th>Unit</th>
-          <th class="right">Unit Price<br><span style="font-size: 6px; font-weight: 500; text-transform: none;">(with tax)</span></th>
-          {tax_header_cells}
-          <th class="right">Amount</th>
-        </tr>
-      </thead>
-      <tbody>
-        {item_rows}
-      </tbody>
-    </table>
-  </section>
-
-  <section class="invoice-sheet__footer">
-    <div class="invoice-sheet__totals">
-      <p class="eyebrow">Tax breakup</p>
-      <p>Taxable: {_fmt_currency(float(invoice.taxable_amount or 0), currency)}</p>
-      {tax_breakup_rows}
-      <p>Total tax: {_fmt_currency(float(invoice.total_tax_amount or 0), currency)}</p>
-      {round_off_html}
-      <p class="eyebrow" style="margin-top: 10px;">Total due</p>
-      <p class="invoice-sheet__total-value">{_fmt_currency(float(invoice.total_amount), currency)}</p>
-      <p class="invoice-amount-words">{_amount_in_words_indian(float(invoice.total_amount), currency)}</p>
-      <p class="muted-text">Authorized by {_e(invoice.company_name) or 'Billing company'}</p>
-    </div>
-  </section>
-
-  <section class="invoice-sheet__bank-section">
-    <p class="eyebrow">Payment details</p>
-    {payment_details_html}
-  </section>
-</div>
-</body>
-</html>"""
-    return html
-
-
-def _build_multi_copy_invoice_html(invoice: Invoice, products: list[Product], invoice_bank_accounts: list[CompanyAccount], copies: int) -> str:
-    if invoice.voucher_type == "purchase":
-        return _build_purchase_invoice_html(invoice, products)
-    if copies == 1:
-        return _build_invoice_html(invoice, products, invoice_bank_accounts, copy_label=_copy_label(1))
-
-    pages = []
-    first_html: str | None = None
-    for i in range(1, copies + 1):
-        full_html = _build_invoice_html(invoice, products, invoice_bank_accounts, copy_label=_copy_label(i))
-        if i == 1:
-            first_html = full_html
-        body_open_end = full_html.index('<body>') + len('<body>')
-        body_close_start = full_html.rindex('</body>')
-        body_content = full_html[body_open_end:body_close_start].strip()
-        page_style = '' if i == copies else ' style="break-after: page; page-break-after: always;"'
-        pages.append(f'<div{page_style}>\n{body_content}\n</div>')
-
-    assert first_html is not None
-    body_open_end = first_html.index('<body>') + len('<body>')
-    combined = '\n'.join(pages)
-    return first_html[:body_open_end] + '\n' + combined + '\n</body>\n</html>'
 
 
 def _build_invoice_pdf(invoice: Invoice, products: list[Product], invoice_bank_accounts: list[CompanyAccount]) -> BytesIO:
