+++ backend/src/services/pdf_templates/__init__.py
@@ -1,0 +1,41 @@
+"""PDF template generation modules for invoices and purchase orders."""
+
+from .builders import (
+    _amount_in_words_indian,
+    _build_pdf_payment_details_html,
+    _build_pdf_table_colgroup,
+    _build_pdf_tax_breakup_rows,
+    _build_pdf_tax_header_cells,
+    _build_pdf_tax_row_cells,
+    _e,
+    _fmt_currency,
+    _fmt_rate,
+    _pdf_display_quantity,
+    _pdf_display_unit,
+    _pdf_unit_price_including_tax,
+)
+from .invoice_template import (
+    _build_invoice_html,
+    _build_multi_copy_invoice_html,
+    _copy_label,
+)
+from .purchase_template import _build_purchase_invoice_html
+
+__all__ = [
+    "_amount_in_words_indian",
+    "_build_invoice_html",
+    "_build_multi_copy_invoice_html",
+    "_build_pdf_payment_details_html",
+    "_build_pdf_table_colgroup",
+    "_build_pdf_tax_breakup_rows",
+    "_build_pdf_tax_header_cells",
+    "_build_pdf_tax_row_cells",
+    "_build_purchase_invoice_html",
+    "_copy_label",
+    "_e",
+    "_fmt_currency",
+    "_fmt_rate",
+    "_pdf_display_quantity",
+    "_pdf_display_unit",
+    "_pdf_unit_price_including_tax",
+]
